Add HTML table rows separation and collection
Once the HTML code is downloaded the function will pass an io.Reader to
the DOM parser, this will detect the beginning of each relevant table
row and keep the children nodes in an independent index of a string list
which will be later accessed by the HTML code analyzer. Notice that both
the opening and closing table row HTML tags will be excluded from the
list as they are not necessary for the analyzys of the code.