Add function to scrape the IPv4 network address
The IP address is obfuscated using CSS code, each section of the address
is wrapped by either a HTML Span or Div tag which can contain a class or
the style attribute. If the HTML tag contains a style attribute with the
display property and value "none" we can consider the whole tag and its
text as invisible. If the tag contains a CSS class and the definition of
the class contains a "display: none" we can consider it invisible too,
and for the remaining text we must strip the HTML code to leave the
bytes that represent the address.