Add function to scrape the location of the server
Each proxy must contains information about its physical location on
Earth, even if the server is part of a virtual network the parent will
generally point to a physical machine located somewhere in the world.
This piece of data contains the name of a country that is believed to be
the real location of the server.