Add function to scrape the proxy connection time
The connection time of each proxy server is represented in percentage to
facilitate the selection of the best network from the list by visually
scanning the table. Notice that the highest the percentage the lowest
the anonimity as some of these servers are controlled by malicious
people or security researchers using these machines as honeypot.