Add function to scrape the approximated proxy speed
The speed of each proxy server is represented in percentages, the
highest the number the better the connection. This is useful if the user
wants to use a proxy with enough stability for certain operations. Note
that the highest the speed the lowest the anonimity as some of these
servers are controlled by malicious users or configured as honeypot.