Add function to scrape the server anonimity level
Some if not all the proxy servers found in the wild are owned by
malicious people, security researchers, or universities. Depending on
the owner we can assume some level of anonimity. Using the servers with
higher anonimity levels is encouraged but be aware that this information
can be faked and you might fall in the hands of a malicious network.