Fix HTTP request without support for Gzip compression
The main service used by this program supports Gzip compression but the
program in itself was not designed for this task. This commit deletes
the HTTP header that sends the list of acceptable encoding, this is
because the list was using Gzip, Deflate, and SDCH but in fact the
program expects string based data.