Fix regular expression for table row discovery
The regular expression used to find each table row in the website has a
bug that forces the program to analyze only half of the entries because
the other half is missing one of the capsules of the RegExp. This commit
fixes this bug and also adds a test file to make sure that this specific
part of the code keeps working.