Add Node.JS parser to help during the debug process
The parsing of the HTML used to scrape the network address is difficult,
I will use a JavaScript parser with a jQuery like module named Cheerio
to select the data that we specifically need to retrieve using the Go
program. This information will be useful during the definition of the
automated tests and maybe in the future for troubleshooting.