Fix regular expressions and missing underscore char
The CSS class name for the HTML tags that will force the invisibility in
some of the strings in the network address table cell contain alphabetic
characters in lower and upper case, numers, underscores and they are
limited to four characters per name. This commit fixes some regular
expressions that were not taking in consideration this condition, as
well as some HTML closing tags.