Add function to concat the protocol, address, and port
It is easier for the user to double click the Terminal in order to copy
the full URL (protocol + IP address + port number) than to copy the
protocol and the IP address in two different actions. Also, the protocol
is currently printed in upper case with the exception of Socks, but the
last one is also unusable because it contains the version supported by
the server. This commit concatenates these three fields for readability.