Modify recurrent code to analyze the HTML table rows
All the export functions process the data from the HTTP response in a
similar way, they iterate over the list of extracted HTML table rows and
then call another function to parse and return a Proxy structure. This
can be simplified if a slice of proxy entries is passed to the export
functions and will also allow to pre-process the data.