Modify padding and width of some columns in table
Both the Speed and Connection columns in the table renderer display a
percentage value, these numbers have better visibility if they are
aligned to the right rather than the left as they are now. The Active
column is also modified to increase the width of the text and the
padding was moved to the right to align the text to the left.