Add generic sorting mechanism for future reuse
This future proof the sorting option for other columns in the slice of
data, for each supported filter the program will set the value of the
proxy Filter property as an integer representing the order of its data
with respect of the other entries in the list. For example, when the
user decides to sort by the speed the program converts the percentage
values to integers and the sort.Sort interface will do the rest.